major update
fixed a bug arising when the network output categories do not match the actual patient categories (e.g. deploying a network to a patient cohort with fewer categories than the training set).
also, some minor improvements to the code which should not change the behavior
tested on 2-class and 3-class datasets